+++ src/neo4j_graphrag/llm/__init__.py
@@ -14,12 +14,14 @@
 #  limitations under the License.
 from .anthropic_llm import AnthropicLLM
 from .base import LLMInterface
+from .cohere_llm import CohereLLM
 from .openai_llm import AzureOpenAILLM, OpenAILLM
 from .types import LLMResponse
 from .vertexai_llm import VertexAILLM
 
 __all__ = [
     "AnthropicLLM",
+    "CohereLLM",
     "LLMResponse",
     "LLMInterface",
     "OpenAILLM",