+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Introduced Vertex AI LLM class for integrating Vertex AI models.
 - Added unit tests for the Vertex AI LLM class.
 - Added support for Cohere LLM and embeddings - added optional dependency to `cohere`.
+- Added support for Anthropic LLM - added optional dependency to `anthropic`.
 
 ### Fixed
 - Resolved import issue with the Vertex AI Embeddings class.